non-gui progress bar methods moved from utility/gui/progress.py to utility/progress.py
